Moved away tests for atomic strategies from Tickets
