+++ tests/Doctrine/ODM/MongoDB/Tests/Functional/AtomicSetTest.php
@@ -5,6 +5,10 @@ namespace Doctrine\ODM\MongoDB\Tests\Functional;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
 use Doctrine\ODM\MongoDB\Tests\QueryLogger;
+use Documents\Book;
+use Documents\Chapter;
+use Documents\IdentifiedChapter;
+use Documents\Page;
 use Documents\Phonebook;
 use Documents\Phonenumber;
 
@@ -390,6 +394,147 @@ class AtomicSetTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
         $this->dm->flush();
         $this->assertCount(1, $this->ql, 'Clearing atomic reference many requires one query');
         $this->dm->clear();
+    }
+
+    public function testAtomicSetUpdatesAllNestedCollectionsInOneQuery()
+    {
+        // Create a book which has one chapter with one page.
+        $chapter1 = new Chapter();
+        $chapter1->pages->add(new Page(1));
+        $book = new Book('title');
+        $book->chapters->add($chapter1);
+
+        $this->dm->persist($book);
+
+        // Saving this book should result in 1 query since we use strategy="atomicSet"
+        $this->dm->flush();
+        $this->assertCount(1, $this->ql, 'Inserting a book with one chapter and page requires one query');
+
+        // Simulate another PHP request which loads this record and tries to add an embedded document two levels deep...
+        $this->dm->clear();
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+
+        // Now we add a new "page" to the only chapter in this book.
+        $firstChapter = $book->chapters->first();
+        $firstChapter->pages->add(new Page(2));
+
+        // Updating this book should result in 1 query since we use strategy="atomicSet"
+        $this->ql->clear();
+        $this->dm->flush();
+        $this->assertCount(1, $this->ql, 'Adding a page to the first chapter of the book requires one query');
+
+        // this is failing if lifecycle callback postUpdate is recomputing change set
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+        $this->assertEquals(2, $book->chapters->first()->pages->count(), "Two page objects are expected in the first chapter of the book.");
+    }
+
+    public function testReplacementOfEmbedManyElements()
+    {
+        // Create a book with a single chapter.
+        $book = new Book();
+        $book->chapters->add(new Chapter('A'));
+
+        // Save it.
+        $this->dm->persist($book);
+        $this->dm->flush();
+
+        // Simulate another PHP request which loads this record.
+        $this->dm->clear();
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+
+        $firstChapter = $book->chapters->first();
+        $firstChapter->name = "First chapter A";
+
+        // Developers commonly attempt to replace the contents of an EmbedMany with a new ArrayCollection like this:
+        $replacementChapters = new ArrayCollection();
+        $replacementChapters->add($firstChapter);
+        $replacementChapters->add(new Chapter('Second chapter B'));
+        $book->chapters = $replacementChapters;
+
+        $this->dm->flush(); // <- Currently getting "Cannot update 'chapters' and 'chapters' at the same time" failures.
+
+        // Simulate another PHP request.
+        $this->dm->clear();
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+
+        // Verify we see chapters A and B.
+        $this->assertEquals('First chapter A', $book->chapters[0]->name);
+        $this->assertEquals('Second chapter B', $book->chapters[1]->name);
+    }
+
+    public function testReplacementOfIdentifiedEmbedManyElements()
+    {
+        $book = new Book();
+        $book->identifiedChapters->add(new IdentifiedChapter('A'));
+
+        $this->dm->persist($book);
+        $this->dm->flush();
+        $this->dm->clear();
+
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+        $firstChapter = $book->identifiedChapters->first();
+        $firstChapter->name = "First chapter A";
+        $replacementChapters = new ArrayCollection();
+        $replacementChapters->add($firstChapter);
+        $replacementChapters->add(new IdentifiedChapter('Second chapter B'));
+        $book->identifiedChapters = $replacementChapters;
+
+        $this->dm->flush();
+        $this->dm->clear();
+
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+        $this->assertEquals('First chapter A', $book->identifiedChapters[0]->name);
+        $this->assertEquals('Second chapter B', $book->identifiedChapters[1]->name);
+    }
+
+    public function testOnlyEmbeddedDocumentUpdated()
+    {
+        // Create a book with a single chapter.
+        $book = new Book();
+        $book->chapters->add(new Chapter('A'));
+
+        // Save it.
+        $this->dm->persist($book);
+        $this->dm->flush();
+
+        // Simulate another PHP request which loads this record.
+        $this->dm->clear();
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+
+        // Modify the chapter's name.
+        $book->chapters->first()->name = "First chapter A";
+
+        $this->dm->flush();
+
+        // Simulate another PHP request & verify the change was saved.
+        $this->dm->clear();
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+
+        $this->assertEquals('First chapter A', $book->chapters[0]->name, "The chapter title failed to update.");
+    }
+
+    public function testUpdatedEmbeddedDocumentAndDirtyCollectionInside()
+    {
+        $book = new Book();
+        $book->chapters->add(new Chapter('A'));
+
+        $this->dm->persist($book);
+        $this->dm->flush();
+        $this->dm->clear();
+
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+        $firstChapter = $book->chapters->first();
+        $firstChapter->name = "Apple";
+
+        // Add some pages.
+        $firstChapter->pages->add(new Page(1));
+        $firstChapter->pages->add(new Page(2));
+
+        $this->dm->flush();
+        $this->dm->clear();
+
+        $book = $this->dm->getRepository(Book::CLASSNAME)->findOneBy(array('_id' => $book->id));
+        $this->assertEquals(2, $book->chapters->first()->pages->count());
     }
 }
